requirements.txt
langchain-openai
mcp-server-fetch
mcp-use
python-dotenv
streamlit
streamlit-extras
